# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "service for distributing test files among processes/machines/containers",
        "title": "Split specs API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "ovr.shevtsov@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "parallel-specs.shelex.dev",
    "basePath": "/",
    "paths": {
        "/api/auth": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "get authorization token",
                "parameters": [
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token response",
                        "schema": {
                            "$ref": "#/definitions/controllers.tokenResponse"
                        }
                    }
                }
            }
        },
...
        "/api/new-password": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "change password for the account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.PasswordChange"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "register new account",
                "parameters": [
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token response",
                        "schema": {
                            "$ref": "#/definitions/controllers.tokenResponse"
                        }
                    }
                }
            }
        },
        "/api/session": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "add new session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.SessionInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "session created",
                        "schema": {
                            "$ref": "#/definitions/controllers.AddSessionResponse"
                        }
                    }
                }
            }
        },
        "/api/session/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "spec"
                ],
                "summary": "get spec executions by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "spec id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pagination",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "session",
                        "schema": {
                            "$ref": "#/definitions/entities.Session"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "delete session by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "session id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/session/{id}/next": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "get next spec file to be executed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "session id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "specify machine id",
                        "name": "machineId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "specify status of previous spec execution",
                        "name": "previousStatus",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "next",
                        "schema": {
                            "$ref": "#/definitions/controllers.NextSpecResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.AddSessionResponse": {
            "type": "object",
            "properties": {
                "projectId": {
                    "type": "string"
                },
                "projectName": {
                    "type": "string"
                },
                "sessionId": {
                    "type": "string"
                }
            }
        },
...
        "controllers.NextSpecResponse": {
            "type": "object",
            "properties": {
                "next": {
                    "type": "string"
                }
            }
        },
        "controllers.PasswordChange": {
            "type": "object",
            "required": [
                "currentPassword",
                "newPassword"
            ],
            "properties": {
                "currentPassword": {
                    "type": "string"
                },
                "newPassword": {
                    "type": "string"
                }
            }
        },
...
        "controllers.SessionInput": {
            "type": "object",
            "required": [
                "projectName",
                "specFiles"
            ],
            "properties": {
                "projectName": {
                    "type": "string"
                },
                "specFiles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "controllers.tokenResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "entities.Execution": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "estimatedDuration": {
                    "type": "integer"
                },
                "finishedAt": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "machineId": {
                    "type": "string"
                },
                "sessionId": {
                    "type": "string"
                },
                "specId": {
                    "type": "string"
                },
                "specName": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "entities.Session": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "integer"
                },
                "executions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Execution"
                    }
                },
                "finishedAt": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "projectId": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "integer"
                }
            }
        },
        "users.User": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 6
                },
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 3
                }
            }
        }
    }
}